# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Food Recognition API",
    "description": "API documentation for your Flask application",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/": {
      "get": {
        "summary": "Hello World",
        "description": "Simple hello world endpoint",
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/logintoken": {
      "post": {
        "summary": "Create Token",
        "description": "Creates an access token for authentication",
        "tags": ["Authentication"],
        "parameters": [
          {
            "name": "email",
            "in": "body",
            "description": "User's email",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "password",
            "in": "body",
            "description": "User's password",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful login",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "access_token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/signup": {
      "post": {
        "summary": "Signup User",
        "description": "Creates a new user",
        "tags": ["User"],
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "description": "Name of the user",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "email",
            "in": "body",
            "description": "Email of the user",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "password",
            "in": "body",
            "description": "Password of the user",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "about",
            "in": "body",
            "description": "About information of the user",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "about": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful signup",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "about": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Email already exists",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "summary": "Logout",
        "description": "Logs out the current user",
        "responses": {
          "200": {
            "description": "Successful logout",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/profile/{getemail}": {
      "get": {
        "summary": "Get User Profile",
        "description": "Retrieves the profile of the specified user",
        "tags": ["User"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "getemail",
            "in": "path",
            "description": "Email of the user whose profile to retrieve",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of user profile",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "about": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/upload_image": {
      "post": {
        "summary": "Upload Image",
        "description": "Uploads an image for object detection",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "image",
            "in": "formData",
            "description": "Image file to upload",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful image upload and object detection",
            "schema": {
              "type": "object",
              "properties": {
                "best_class_name": {
                  "type": "string"
                },
                "highest_confidence": {
                  "type": "number",
                  "format": "float"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (no file part, no selected file, or invalid file type)",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}